Provide matrix tile in separate in examples
